django-framework/requirements.txt
asgiref==3.4.1
Django==3.2.8
pytz==2021.3
sqlparse==0.4.2
psycopg2-binary==2.9.3
django-crispy-forms==1.14.0
django-simple-history==3.0.0
Gunicorn==20.1.0
whitenoise==5.3.0
matplotlib==3.1.2